+++ src/components/reflection-all/ReflectionAllArea.tsx
@@ -1,63 +1,38 @@
 import Grid from "@mui/material/Grid2";
 import { ReflectionWithUser } from "@/src/api/reflection-api";
 import ReflectionCardWithIcon from "./ReflectionCardWithIcon";
-import { Box, keyframes, Typography, useMediaQuery } from "@mui/material";
+import { Box, Typography, useMediaQuery } from "@mui/material";
 import { theme } from "@/src/utils/theme";
 import ToOtherPageButton from "./ToOtherPageButton";
 import { User } from "@prisma/client";
-import ArrowOnlyPagination from "@/src/components/pagination/ArrowPagination";
 import { ChangeEvent } from "react";
 import { animation } from "../shared/animation";
+import ArrowPagination from "@/src/components/pagination/ArrowPagination";
+import ReflectionAllHeader from "../pagination/ReflectionAllHeader";
 
 type ReflectionAllAreaProps = {
-  reflections: ReflectionWithUser[];
   currentUsername: User["username"];
+  reflections: ReflectionWithUser[];
   currentPage: number;
   totalPage: number;
   handleChange: (event: ChangeEvent<unknown>, value: number) => void;
 };
 
 const ReflectionAllArea: React.FC<ReflectionAllAreaProps> = ({
-  reflections,
   currentUsername,
+  reflections,
   currentPage,
   totalPage,
   handleChange,
 }) => {
-  const isSmallScreen = useMediaQuery(theme.breakpoints.down("md"));
-
-  const Header = () => (
-    <>
-      {isSmallScreen && (
-        <Box position={"absolute"} top={-45} right={30}>
-          <ToOtherPageButton currentUsername={currentUsername} />
-        </Box>
-      )}
-      <Box
-        display={"flex"}
-        justifyContent={"space-between"}
-        mx={{ xs: 4, md: 3 }}
-      >
-        <Typography component={"h1"} fontSize={17} letterSpacing={1}>
-          みんなの振り返り(公開のみ)
-        </Typography>
-        {!isSmallScreen && (
-          <ToOtherPageButton currentUsername={currentUsername} />
-        )}
-      </Box>
-    </>
-  );
-
   return (
     <Box mt={12} position={"relative"}>
-      <Header />
-      <Box position={"absolute"} mt={2} right={18}>
-        <ArrowOnlyPagination
-          currentPage={currentPage}
-          totalPage={totalPage}
-          onChange={handleChange}
-        />
-      </Box>
+      <ReflectionAllHeader
+        currentUsername={currentUsername}
+        currentPage={currentPage}
+        totalPage={totalPage}
+        handleChange={handleChange}
+      />
       <Grid container my={{ xs: 4, md: 8 }}>
         {reflections.map((reflection, index) => (
           <Grid
